refactor(colorbar): simplify ViewportColorbar and ViewportColorbarsContainer components by removing unused state and close button logic
